added animation and adjusted styling of nav menu
