+++ package.json
@@ -14,7 +14,6 @@
 		"@emotion/react": "^11.14.0",
 		"emotion": "^11.0.0",
 		"js-cookie": "^3.0.5",
-		"lucide-react": "^0.544.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-router-dom": "^6.22.3",
