+++ src/components/Pages/Home.styles.ts
@@ -196,59 +196,3 @@ export const aboutMeTextStyle: CSSObject = {
 	marginBottom: 0,
 	marginTop: '0.6rem',
 };
-
-export const hamburgerMenuStyle: CSSObject = {
-	position: 'fixed',
-	top: '1rem',
-	right: '1rem',
-	zIndex: 1000,
-};
-
-export const hamburgerButtonStyle: CSSObject = {
-	background: 'rgba(255, 255, 255, 0.09)',
-	border: 'none',
-	borderRadius: '8px',
-	padding: '0.75rem',
-	color: '#B6ADFF',
-	backdropFilter: 'blur(10px)',
-	boxShadow: '0 4px 6px rgba(0, 0, 0, 0.1)',
-	transition: 'all 0.2s ease-in-out',
-	display: 'flex',
-	alignItems: 'center',
-	justifyContent: 'center',
-	'&:hover': {
-		background: 'rgba(255, 255, 255, 0.2)',
-	},
-	'&:active': {
-		background: 'rgba(255, 255, 255, 0.2)',
-	},
-};
-
-export const menuDropdownStyle: CSSObject = {
-	position: 'absolute',
-	top: '100%',
-	right: 0,
-	marginTop: '0.5rem',
-	background: 'rgba(255, 255, 255, 0.1)',
-	backdropFilter: 'blur(10px)',
-	border: '1px solid rgba(255, 255, 255, 0.2)',
-	borderRadius: '8px',
-	padding: '0.5rem 0',
-	minWidth: '150px',
-	boxShadow: '0 8px 16px rgba(0, 0, 0, 0.2)',
-	animation: 'fadeInDown 0.2s ease-out',
-};
-
-export const menuItemStyle: CSSObject = {
-	display: 'block',
-	padding: '0.75rem 1rem',
-	color: '#B6ADFF',
-	textDecoration: 'none',
-	fontSize: '0.9rem',
-	fontWeight: 500,
-	transition: 'all 0.2s ease-in-out',
-	'&:hover': {
-		background: 'rgba(255, 255, 255, 0.1)',
-		color: '#FFFFFF',
-	},
-};
